requirements.txt
accelerate==1.10.1
addict==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
av==15.1.0
blinker==1.9.0
brotlicffi @ file:///croot/brotlicffi_1736182461069/work
certifi @ file:///croot/certifi_1754570635119/work/certifi
cffi @ file:///croot/cffi_1736182485317/work
charset-normalizer==3.4.3
chumpy==0.70
click==8.2.1
comm==0.2.3
ConfigArgParse==1.7.1
contourpy==1.3.2
cycler==0.12.1
dash==3.2.0
decorator==5.2.1
exceptiongroup==1.3.0
executing==2.2.1
fastjsonschema==2.21.2
filelock @ file:///croot/filelock_1744281381737/work
Flask==3.1.2
fonttools==4.59.2
freetype-py==2.5.1
frozenlist==1.7.0
fsspec==2025.9.0
gmpy2 @ file:///croot/gmpy2_1738085463648/work
hf-xet==1.1.10
huggingface-hub==0.35.0
idna==3.10
imageio==2.37.0
importlib_metadata==8.7.0
ipython==8.37.0
ipywidgets==8.1.7
itsdangerous==2.2.0
jedi==0.19.2
Jinja2 @ file:///croot/jinja2_1741710844255/work
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_core==5.8.1
jupyterlab_widgets==3.0.15
kiwisolver==1.4.9
manotorch @ git+https://github.com/lixiny/manotorch.git@2f6a701e76ee544bbeac1d3e72c628061f585a6c
MarkupSafe @ file:///croot/markupsafe_1738584038848/work
matplotlib==3.10.6
matplotlib-inline==0.1.7
mkl-service==2.4.0
mkl_fft @ file:///io/mkl313/mkl_fft_1730824109137/work
mkl_random @ file:///io/mkl313/mkl_random_1730823916628/work
mpmath @ file:///croot/mpmath_1690848262763/work
multidict==6.6.4
narwhals==2.5.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx @ file:///croot/networkx_1737039604450/work
numpy==1.23.5
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
open3d==0.19.0
opencv-python==4.11.0.86
packaging==25.0
pandas==2.3.2
parso==0.8.5
pexpect==4.9.0
pillow @ file:///croot/pillow_1752524526621/work
platformdirs==4.4.0
plotly==6.3.0
prompt_toolkit==3.0.52
propcache==0.3.2
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser @ file:///home/task_175749556993857/conda-bld/pycparser_1757496041120/work
pyg-lib==0.4.0+pt25cu118
pyglet==2.1.8
Pygments==2.19.2
PyOpenGL==3.1.0
pyparsing==3.2.4
pyquaternion==0.9.9
pyrender==0.1.45
PySocks @ file:///home/builder/ci_310/pysocks_1640793678128/work
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML @ file:///croot/pyyaml_1728657952215/work
qwen-vl-utils==0.0.11
referencing==0.36.2
regex==2025.9.1
requests @ file:///croot/requests_1756709366904/work
retrying==1.4.2
rpds-py==0.27.1
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.15.3
six==1.17.0
stack-data==0.6.3
sympy @ file:///croot/sympy_1756713465822/work
threadpoolctl==3.6.0
tokenizers==0.19.1
torch==2.5.1
torch-geometric==2.6.1
torch_cluster==1.6.3+pt25cu118
torch_scatter==2.1.2+pt25cu118
torch_sparse==0.6.18+pt25cu118
torch_spline_conv==1.2.2+pt25cu118
torchaudio==2.8.0+cpu
torchvision==0.20.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.44.0
trimesh==4.8.1
triton==3.1.0
typing_extensions @ file:///croot/typing_extensions_1756280817316/work
tzdata==2025.2
urllib3 @ file:///croot/urllib3_1750775463400/work
wcwidth==0.2.13
Werkzeug==3.1.3
widgetsnbextension==4.0.14
yarl==1.20.1
zipp==3.23.0
